backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Deep Learning
torch==2.0.0
torchvision==0.15.0
timm==0.9.12
huggingface-hub==0.19.4

# Image processing
pillow==10.1.0
opencv-python-headless==4.8.1.78
scikit-image==0.22.0

# Data processing
numpy>=1.26.0
scipy==1.11.4
scikit-learn==1.3.2

# Utilities
tqdm==4.66.1
